requirements.txt
torch>=1.8.0
torchvision>=0.9.0
torch-geometric>=2.0.0
dgl>=0.8.0
dgllife>=0.2.8
rdkit-pypi>=2022.3.5
transformers>=4.18.0
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.62.0
prefetch-generator>=1.0.1
biopython>=1.79
deepchem>=2.6.0
networkx>=2.6
matplotlib>=3.5.0
seaborn>=0.11.0 